move lock state to LockCore
