+++ CounterApp/Lock/LockView.swift
@@ -8,34 +8,6 @@
 import ComposableArchitecture
 import SwiftUI
 
-struct LockState: Equatable {
-    var counters: IdentifiedArrayOf<CounterState> = []
-    var isPresent: Bool = false
-}
-
-enum LockAction: Equatable {
-    case counter(id: CounterState.ID, action: CounterAction)
-    case setSheet(isPresented: Bool)
-}
-
-struct LockEnvironment {}
-
-let lockReducer: Reducer<LockState, LockAction, LockEnvironment> =
-    counterReducer.forEach(
-        state: \.counters,
-        action: /LockAction.counter(id:action:),
-        environment: { _ in CounterEnviroment.defaultEnv() }
-    ).combined(with: Reducer { state, action, _ in
-        switch action {
-        case let .setSheet(isPresented):
-            state.isPresent = isPresented
-            return .none
-
-        default:
-            return .none
-        }
-    })
-
 struct LockView: View {
     let store: Store<LockState, LockAction>
